refactor SearchBar and implement UseMemoCard for character search functionality
