+++ src/App.tsx
@@ -2,8 +2,9 @@ import { Card, CardContent } from '@/components/ui/card';
 import '@/public/styles/globals.css';
 // import { Counter, CounterDisplay } from './components/shared';
 // import { useState } from 'react';
-import { ProfileCard,DragonBallCharacterCard, SearchBar } from './components/shared';
+import { ProfileCard,DragonBallCharacterCard} from './components/shared';
 import { profileData } from './data';
+import { UseMemoCard } from './hooks';
 
 
 export function App() {
@@ -30,7 +31,7 @@ export function App() {
         </CardContent>
       </Card> */}
       {/* <ProfileCard src={profileData.image} name={profileData.name} description={profileData.description} linkedin={profileData.socialMedia[1].url} github={profileData.socialMedia[0].url} /> */}
-      <SearchBar />
+      <UseMemoCard />
     </div>
   );
 }
